# perf: namedtuples are hashable, don't need a key