`namedtuple` are like Gondor, they need no key.

The intention is to use this as a key for the LazyCache, so it needs to be hashable. As `namedtuple`, `dtype` is already hashable.

If the `ShapeTracker` where a tuple of views, it would be hashable by itself as well.


`python3.11 -O test/external/external_test_speed_llama.py`
Before

```
codegen         mean runtime:  193.99ms, runs:   229.83,  246.88,  180.52,  176.26,  195.84,  178.55,  184.07,  187.23,  176.50,  184.28
methodcache     mean runtime:  175.20ms, runs:   209.10,  197.53,  162.70,  161.03,  162.69,  188.16,  163.63,  173.66,  167.63,  165.84
profile         mean runtime:  830.36ms, runs:   807.96,  892.16,  756.62,  899.99,  830.83,  803.15,  790.21,  789.10,  852.20,  881.34
```


After
```
codegen         mean runtime:  188.02ms, runs:   185.72,  238.30,  193.68,  176.52,  176.44,  176.48,  181.97,  189.14,  175.11,  186.82
methodcache     mean runtime:  169.99ms, runs:   158.11,  225.25,  166.17,  178.28,  161.48,  167.16,  161.11,  159.78,  161.47,  161.11
profile         mean runtime:  788.67ms, runs:   770.87,  798.12,  804.25,  781.07,  776.63,  750.59,  799.22,  801.23,  775.16,  829.60
```