@tinyb0t 2023-08-21
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/helpers.py                        0               2
tinygrad/lazy.py                           2               2
tinygrad/runtime/lib.py                    2               2
------------------------------------------------------------
total                                      4               6
------------------------------------------------------------
lines added in the tinygrad folder: -2
```